Finish implemnting backend stuff for seller

+++ src/main/java/com/f23/shoppeasy/model/listing/Listing.java
@@ -27,8 +27,8 @@ public class Listing {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
     
-    private int price;
-    private int discount;
+    private double price;
+    private double discount;
     private int sellerId;
     private String name;
     private String deliveryOrigin;
